+++ types-scripts/native-modules.ts
@@ -96,3 +96,11 @@ export import TextFieldFormElement = formElement.TextFieldFormElement;
 //@ts-ignore
 import forms = require('../src/forms/Forms');
 export import Forms = forms.Forms;
+
+//@ts-ignore
+import aiConfig = require('../src/configuration/AIAssistantConfiguration');
+export import AIAssistantConfiguration = aiConfig.AIAssistantConfiguration;
+
+//@ts-ignore
+import pageInfo = require('../src/document/PDFPageInfo');
+export import PDFPageInfo = pageInfo.PDFPageInfo;